Move most inifile tests to flit_core
